Moved io.py inside data_factories/
